fix: resolve all CI failures — fmt, clippy, macOS unused imports
- cargo fmt --all: fix formatting across all crates
- clippy --workspace -D warnings: fix all lints
  - collapsible if statements (ciri-anim, ciri-input, ciri-render)
  - let...else → ? operator (ciri/ui.rs)
  - loop → while let (ciri/connection.rs)
  - if let Err(_) → .is_err() (ciri/connection.rs)
  - boolean simplification (input_handler.rs, mouse.rs)
  - identical if blocks (ui.rs tab_text_color)
  - too_many_arguments allow (render.rs, status_bar.rs)
  - empty doc comment line (paste_guard.rs)
  - SpringParams::default → impl Default (ciri-anim/spring.rs)
  - impl Default for Osc7Parser (ciri-term)
- macOS build: gate GL-only imports/constants behind cfg(not(macos))
  in ciri-gpu/gl.rs (ConfigTemplateBuilder, ContextApi, SwapInterval,
  HasDisplayHandle, RECT_VS, RECT_FS, GlRectPipeline::new,
  GlyphAtlasGpu::new)

+++ crates/ciri/src/app/status_bar.rs
@@ -1,5 +1,6 @@
 use ciri_render::glyph_cache::{GlyphCache, GlyphInstance};
 
+#[allow(clippy::too_many_arguments)]
 pub(crate) fn emit_status_text(
     atlas: &mut GlyphCache,
     text: &str,
@@ -15,8 +16,8 @@ pub(crate) fn emit_status_text(
             && entry.width > 0
             && entry.height > 0
         {
-            let sx = x_start + i as f32 * cell_width + entry.bearing_x as f32;
-            let sy = text_y + baseline - entry.bearing_y as f32;
+            let sx = x_start + i as f32 * cell_width + entry.bearing_x;
+            let sy = text_y + baseline - entry.bearing_y;
             glyphs.push(GlyphInstance {
                 pos: [sx, sy],
                 size: [entry.width as f32, entry.height as f32],
